moved triggerBroadcast method to api
