+++ lib/api/requests.js
@@ -1,5 +1,3 @@
-const fs = require('fs');
-
 const REQUIRED_FIELDS = Object.freeze(['to', 'identifiers']);
 const OPTIONAL_FIELDS = Object.freeze([
   'transactional_message_id',
