+++ lib/track.js
@@ -2,22 +2,6 @@ const Request = require('./request');
 const { Region, RegionUS } = require('./regions');
 const { isEmpty } = require('./utils');
 
-const BROADCASTS_ALLOWED_RECIPIENT_FIELDS = {
-  ids: ['ids', 'id_ignore_missing'],
-  emails: ['emails', 'email_ignore_missing', 'email_add_duplicates'],
-  per_user_data: ['per_user_data', 'id_ignore_missing', 'email_ignore_missing', 'email_add_duplicates'],
-  data_file_url: ['data_file_url', 'id_ignore_missing', 'email_ignore_missing', 'email_add_duplicates'],
-};
-
-const filterRecipientsDataForField = (recipients, field) => {
-  return BROADCASTS_ALLOWED_RECIPIENT_FIELDS[field].reduce((obj, field) => {
-    if (!!recipients[field]) {
-      obj[field] = recipients[field];
-    }
-    return obj;
-  }, {});
-};
-
 class MissingParamError extends Error {
   constructor(param) {
     super(param);
@@ -129,21 +113,5 @@ module.exports = class TrackClient {
     return this.request.destroy(
       `${this.trackRoot}/customers/${encodeURIComponent(customerId)}/devices/${encodeURIComponent(deviceToken)}`,
     );
-  }
-
-  triggerBroadcast(id, data, recipients) {
-    let payload = {};
-    let customRecipientField = Object.keys(BROADCASTS_ALLOWED_RECIPIENT_FIELDS).find((field) => recipients[field]);
-
-    if (customRecipientField) {
-      payload = Object.assign({ data }, filterRecipientsDataForField(recipients, customRecipientField));
-    } else {
-      payload = {
-        data,
-        recipients,
-      };
-    }
-
-    return this.request.post(`${this.apiRoot}/api/campaigns/${id}/triggers`, payload);
   }
 };
